Refactor hexagonal  | moved context history
