+++ src/bot/application/chat_handler.py
@@ -1,5 +1,4 @@
 import os
-import json
 import asyncio
 import logging
 from collections import defaultdict
@@ -8,66 +7,21 @@ from concurrent.futures import ThreadPoolExecutor
 from fastapi import WebSocket, WebSocketDisconnect
 
 from src.bot.domain.chat_bot import ChatBot
+from src.bot.domain.services.chat_context_history_service import (
+    ChatContextHistory,
+    ChatContextHistoryPoolService,
+)
 from src.bot.infrastructure.bots.chat_bot_testings import ChatBotTestingsMock
 from src.bot.infrastructure.bots.ollama_chat_bots import (
     ChatBotOllamaOpenAI,
     ChatBotOllamaNeuralChat,
 )
-from src.cache.domain.cache import Cache
 from src.settings import Settings
 
 
 logger = logging.getLogger(__name__)
 
-from src.cache.application.get_cache import get_cache
-
-
-class ChatHistory:
-    def __init__(self, application_id: str, session_id: str):
-        self._application_id = application_id
-        self._session_id = session_id
-        self._settings: Settings = Settings.get()
-        self._cache: Cache = get_cache()
-
-        self._cache_key = f"{self._settings.cache_service_prefix}{self._application_id}:{self._session_id}"
-
-    async def add_context(self, context: list[dict]) -> None:
-        context_encoded = json.dumps(context)
-        await self._cache.set(self._cache_key, context_encoded, ex=300)
-
-    async def get_context(self) -> list[dict] | None:
-        cached = await self._cache.get(self._cache_key)
-        if cached is None:
-            return None
-        history = json.loads(cached)
-        return history
-
-
-class ChatHistoryPool:
-    def __init__(self):
-        self._sessions: dict[tuple[str, str], ChatHistory] = {}
-
-    def create_history(self, application_id: str, session_id: str) -> ChatHistory:
-        chat_history = ChatHistory(application_id, session_id)
-        self._sessions[(application_id, session_id)] = chat_history
-        return chat_history
-
-    def get_history(self, application_id: str, session_id: str) -> ChatHistory | None:
-        return self._sessions.get((application_id, session_id), None)
-
-    def get_or_create_history(
-        self, application_id: str, session_id: str
-    ) -> ChatHistory:
-        chat_history = self._sessions.get((application_id, session_id), None)
-        if chat_history is None:
-            chat_history = self.create_history(application_id, session_id)
-        return chat_history
-
-    def remove_history(self, application_id: str, session_id: str) -> None:
-        del self._sessions[(application_id, session_id)]
-
-
-chat_history_pool = ChatHistoryPool()
+chat_history_pool = ChatContextHistoryPoolService()
 
 
 class ConnectionManager:
@@ -120,8 +74,8 @@ class ChatHandler:
         self._session_id: str = session_id
         self._websocket: WebSocket = websocket
         self._active_connections: dict[str, list[WebSocket]] = {}
-        self._chat_history: ChatHistory = chat_history_pool.get_or_create_history(
-            application_id, session_id
+        self._chat_history: ChatContextHistory = (
+            chat_history_pool.get_or_create_history(application_id, session_id)
         )
 
     async def handle(self) -> None:
